requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
# backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2
django-allauth==0.50.0
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-modeltranslation==0.17.5
django-ranged-response==0.2.0
django-recaptcha3==0.4.0
django-rest==0.8.7
django-simple-captcha==0.5.17
djangorestframework==3.13.1
filelock==3.6.0
flower
h11==0.13.0
idna==3.3
kombu==5.2.4
lxml==4.8.0
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pbr==5.8.1
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.14.0
wsproto==1.1.0
